Added instrument: Newport Agilis
Added a new instrument for Newport Agilis devices. Currently, only the
AG-UC2 Controller is supported. Added the controller to the respective
ini file in the newport folder.

Axis class implemented to send device specific commands. This can be
used from other controllers, once implemented, as well.

Some caveats / non-implemented and only partially tested routines are
noted on top of the `agilis.py` file. I don't have a device with a limit
switch available for testing, therefore, some routines that only work
on such devices are not implemented and some routines are untested. The
doc strings for every routine states this clearly as well.

A full test suite for all functions is also provided.

Everything testing great with tox for py36, py37, and py38.