Change TekDPO70000 to always read termination character after waveform
The loopback communicator used in testing does not flush the input,
thus a termination character is stuck in the queue that does not get
read. According to the manual, a `<newline>` character is always sent
after transmitting the waveform.
Flushing the input is now simplified and includes a `read_raw` for one
character. This takes care of proper testing. Tests were adjusted
accordingly.