Status word now created according to manual, read in w/o last byte
The status word, according to the manual, contains the termination
character. It is unclear if this is the same as the end line end
character, was implemented that way before.
The test is generating a termination character different from the line
end character (to keep it general).
The unpacking of the status word is limited to the 13 bytes that
we want to unpack.