Test suite and BFs for Keithley580 (#263)
* Test suite and BFs for Keithley580

Test suite with full coverage added

**BFs**:
- Removed inaccessible code block
- Replaced a not in `enum` statement with a check of `isinstance`, now
  the test is compatible to the error description that is raised.
- Getting the 'auto' input was previously not possible, fixed the
  respective property getter.
- Fixed typos
- Fixed status word getting such that an IOError can actually be
  raised now if failing to receive a status word after given number of
  tries
- Fixed reading of bytes: same issue as in Keithley195, trying to
  unpack a string with `struct.unpack`. Now consistent and functional.

* Read measurement and status word using `read_raw`

Preferred method over using `query`, since `query` decodes the return,
which then needs to be encoded again prior to unpacking with `struct`.
Now, the return is read using `read_raw`, which can directly be
unpacked.
Note: `query('')` still sends a `:`, which is now sent using `sendcmd`
prior to running `read_raw`. It also strips the last character from
the return, which done here as well.
Updated tests.

* Status word now created according to manual, read in w/o last byte

The status word, according to the manual, contains the termination
character. It is unclear if this is the same as the end line end
character, was implemented that way before.
The test is generating a termination character different from the line
end character (to keep it general).
The unpacking of the status word is limited to the 13 bytes that
we want to unpack.

* Removed terminator from status word createn in test fixture

The terminator is already added by expected protocol.

Co-authored-by: Steven Casagrande <scasagrande@galvant.ca>