Test suite and BF for Newport's `error.py`
Simple test suite with full coverage. Time stamps are only tested for
existence. Further tests is probably not required, however, if wanted
we might want to use `pytest-freezegun` and include that into the
`dev-requirements.txt` in order to freeze timestamps. (Maybe should
include that just for the awesome name).

Bug fix:
The timedelta creation for the timestamp made absolutely no sense and
resulted in an error, even when just executing it by itself.
Creating a timestamp that is as close to the one given if not specific
`datetime` is given.