Rewrote instrument.py and usb communicator to work with Newport PMC8742
Tests are currently failing, but that's okay, these are tests that
simply test for the old / not-working usb communicator initalization.

SRS830 test is failing because USBCommunicator is used to test
some behavior that USBCommunicator is the wrong one. Anyway, probably
better to just mock something here in this failing test.