Establish read and write capabilities for USBCommunicator (#313)
* Support for Newport Picomotor Controller 8742

Full functionality for this controller in single-controller and
multi-controller mode (main / secondaries via RS-485) is implemented.
Only single-controller mode is tested with hardware at the moment.
Full test suite however includes tests for multi-controller modes.
Routines were documented in docstring and also in Sphinx
documentation.

Note: Control via USB is currently not functional, since
InstrumentKit has some issues with the real USB communications
class. Needs separate PR.

* Added a `read_raw` routine to `instrument.py`

Reading raw data from the instrument is required for the picomotor
control class, however, had so far to be done by directly accessing
a private argument of `instrument.py` (not pretty).

* Fix writing to USB device, change PMC to 1 termination character

Writing works, but has issues with the terminator. need to add it at
least once manually...
The IK way of doing things is currently not the same as the tutorial?
Reading always and only returns 0... might need to switch the control?

* Revert changes to newport_pmc8742 - termination character change

* Rewrote instrument.py and usb communicator to work with Newport PMC8742

Tests are currently failing, but that's okay, these are tests that
simply test for the old / not-working usb communicator initalization.

SRS830 test is failing because USBCommunicator is used to test
some behavior that USBCommunicator is the wrong one. Anyway, probably
better to just mock something here in this failing test.

* Adjust USBCommunicator such that compatible with other instruments

* Fixed failing tests in existing test suite

* Full testing for USBCommunicator, some BFs

* USB Communicator bug fix and test suite...

now for real...

* Use wMaxPacketSize for the size of the package that is read by default