Add 'direction' property for move_indefinitely with a negative velocity.
Negative velocities would throw a MAXIMUM VELOCITY EXCEEDED error, and move_indefinitely only sent the "MV" command. This meant there was no was to move_indefinitely in the 'negative' direction. Now, when setting velocity, a direction will be set as well, allowing for indefinite movement in either direction using the "MV+" or "MV-" commands. This is the only area that direction currently applies.